+++ 3dEngine/src/texture/filter/bilateralblur/BilateralBlurFilterBuilder.cpp
@@ -46,9 +46,9 @@ namespace urchin {
         assert(getDepthTexture());
 
         if (getTextureType() != TextureType::DEFAULT) {
-            throw std::invalid_argument("Unsupported texture type for bilateral blur filter: " + std::to_string(getTextureType()));
+            throw std::invalid_argument("Unsupported texture type for bilateral blur filter: " + std::to_string((int)getTextureType()));
         } else if (getTextureFormat() != TextureFormat::GRAYSCALE_8_INT && getTextureFormat() != TextureFormat::GRAYSCALE_16_FLOAT) {
-            throw std::invalid_argument("Unsupported texture format for bilateral blur filter: " + std::to_string(getTextureFormat()));
+            throw std::invalid_argument("Unsupported texture format for bilateral blur filter: " + std::to_string((int)getTextureFormat()));
         }
 
         std::unique_ptr<TextureFilter> textureFilter;
